GUA-209: Move credential builders out of controller
